Move type resolution files out of utils and into types folder
